Remove unused imports

+++ src/org/hyperic/hq/ui/action/resource/common/inventory/RemoveResourceGroupsAction.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of
  * "derived work".
  * 
- * Copyright (C) [2004, 2005, 2006], Hyperic, Inc.
+ * Copyright (C) [2004-2008], Hyperic, Inc.
  * This file is part of HQ.
  * 
  * HQ is free software; you can redistribute it and/or modify
@@ -23,34 +23,25 @@
  * USA.
  */
 
-/*
- * Created on Mar 19, 2003
- *
- */
 package org.hyperic.hq.ui.action.resource.common.inventory;
 
-import java.io.IOException;
 import java.util.HashMap;
 
 import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import org.hyperic.hq.appdef.shared.AppdefEntityID;
-import org.hyperic.hq.authz.shared.PermissionException;
-import org.hyperic.hq.bizapp.shared.AppdefBoss;
-import org.hyperic.hq.common.ApplicationException;
-import org.hyperic.hq.ui.Constants;
-import org.hyperic.hq.ui.action.BaseAction;
-import org.hyperic.hq.ui.util.ContextUtils;
-import org.hyperic.hq.ui.util.RequestUtils;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
+import org.hyperic.hq.appdef.shared.AppdefEntityID;
+import org.hyperic.hq.bizapp.shared.AppdefBoss;
+import org.hyperic.hq.ui.Constants;
+import org.hyperic.hq.ui.action.BaseAction;
+import org.hyperic.hq.ui.util.ContextUtils;
+import org.hyperic.hq.ui.util.RequestUtils;
 
 /**
  *
@@ -80,7 +71,7 @@ public class RemoveResourceGroupsAction extends BaseAction {
         AppdefBoss boss = ContextUtils.getAppdefBoss(ctx);            
         Integer sessionId = RequestUtils.getSessionId(request);
         AppdefEntityID entityId = new AppdefEntityID(resourceType.intValue(), 
-            resourceId.intValue());
+                                                     resourceId);
 
         Integer[] groups = rmForm.getG();
         if (groups != null) {
